requirements.txt
aiogram>=3.22.0
fastapi>=0.116.0
uvicorn>=0.35.0
python-dotenv>=1.1.1
